Update "weightedRegression.py" at v0.1; Arrange ./Doc/dhdt for v0.1
