--- Doc/dhdt/Hooker/Hooker_p10m10_3columns.tif.aux.xml
@@ -1,20 +1,0 @@
-<PAMDataset>
-  <PAMRasterBand band="1">
-    <Histograms>
-      <HistItem>
-        <HistMin>-10.85064325904846</HistMin>
-        <HistMax>10.91698656654358</HistMax>
-        <BucketCount>1000</BucketCount>
-        <IncludeOutOfRange>0</IncludeOutOfRange>
-        <Approximate>0</Approximate>
-        <HistCounts>1|0|0|0|1|0|0|0|0|0|0|0|0|1|0|1|0|2|2|0|1|0|0|1|1|0|1|3|2|0|2|1|1|1|0|0|1|0|0|18|17|21|22|16|21|30|21|28|18|22|22|22|23|28|29|17|18|16|19|29|12|17|29|21|14|17|17|26|18|23|25|23|20|22|22|17|21|20|16|16|20|23|20|28|32|27|21|28|22|26|33|29|33|23|23|25|21|17|16|18|18|23|21|28|16|19|34|22|29|13|28|19|27|19|29|22|21|26|20|30|22|17|29|19|15|25|25|31|19|23|21|22|19|22|19|31|24|20|28|27|22|22|30|23|40|23|21|28|33|40|23|26|24|20|33|28|28|24|32|24|39|33|23|31|31|19|30|19|22|32|24|30|19|27|36|28|29|33|29|32|30|29|30|25|29|30|30|40|39|43|38|25|26|28|33|49|29|32|27|28|26|36|27|38|37|38|27|26|30|31|25|25|23|34|25|30|22|25|26|25|24|26|24|36|29|28|30|19|26|26|28|34|33|34|26|27|24|38|20|34|28|34|30|32|47|32|25|29|21|30|26|25|29|32|45|35|22|36|31|27|22|36|24|45|29|38|33|28|36|36|35|27|39|36|39|30|26|34|43|42|42|34|38|31|38|29|42|45|30|36|39|27|36|37|38|34|34|51|30|33|37|45|43|43|40|48|32|34|45|38|48|38|46|40|40|35|41|50|41|35|50|46|54|51|41|44|42|39|51|48|44|52|52|44|48|41|56|43|37|48|56|52|58|57|61|55|53|49|64|69|60|65|65|68|71|57|69|80|68|65|61|63|83|84|77|66|80|83|76|87|71|86|85|104|107|111|118|119|129|112|114|146|130|162|131|147|163|157|188|180|207|185|199|194|181|197|216|214|226|196|220|228|231|223|209|217|228|230|206|232|263|268|251|261|267|280|284|290|293|308|332|330|336|386|401|399|431|383|395|443|476|452|491|467|461|514|528|571|533|530|592|548|545|561|649|563|666|663|666|691|760|880|927|1080|1194|1269|1386|1566|1719|1866|2024|2092|2099|2058|2205|2351|2517|2695|2788|2835|2829|2898|3004|2977|3025|2894|2794|2740|2786|2566|2416|2395|2345|2387|2330|2291|2267|2321|2410|2472|2527|2593|2762|2702|2731|2825|2745|2703|19863|2449|2197|1950|1816|1587|1408|1364|1163|1047|878|750|686|646|618|576|535|471|469|444|392|384|350|362|334|309|293|289|244|206|208|217|182|202|174|190|160|184|144|142|141|141|124|127|125|114|118|117|116|117|119|111|92|105|87|99|90|88|86|91|91|85|77|76|76|77|72|86|76|58|61|57|58|55|52|69|57|65|49|60|59|57|60|56|57|53|42|48|44|49|49|58|45|49|52|51|55|61|49|46|60|54|53|64|49|60|35|46|43|41|47|32|47|34|53|40|44|42|37|48|48|42|22|35|38|49|43|44|41|41|22|33|30|37|37|39|35|39|35|35|41|31|34|36|34|35|28|33|28|41|27|35|35|47|37|43|31|32|37|30|44|32|28|33|30|24|29|34|36|37|22|20|25|31|35|19|36|31|39|27|33|29|36|37|23|40|39|35|31|33|27|33|25|34|41|25|30|26|33|25|25|25|26|25|35|25|23|32|23|30|24|22|32|30|23|30|24|34|23|18|22|14|27|22|26|28|16|26|29|23|17|21|21|22|24|21|23|28|21|25|17|23|25|25|22|16|26|17|15|17|18|22|22|27|19|21|26|24|25|20|23|25|24|26|14|21|17|28|21|19|21|21|12|15|20|25|18|16|17|22|21|23|19|23|28|24|19|21|17|20|17|28|14|22|19|24|19|15|23|22|21|21|20|29|19|33|27|16|26|11|24|23|24|12|25|21|23|15|20|15|8|17|16|15|18|14|15|19|22|23|23|13|18|26|16|16|20|21|22|14|19|17|20|17|12|13|12|16|16|10|19|17|15|17|16|17|12|19|12|18|10|17|16|14|14|12|16|13|8|18|10|4|18|11|13|14|12|13|22|15|16|15|10|11|8|10|10|15|10|17|17|6|15|11|15|11|21|12|7|9|7|9|9|9|13|7|12|11|12|10|15|11|7|12|14|14|7|13|18|8|11|13|12|9|7|13|17|13|11|18|9|13|19|9|12|10|16|12|13|24|11|6|16|15|13|6|7|11|11|12|12|10|10|8|5|8|6|6|8|6|1|0|0|0|0|0|0|0|0|0|1|0|0|0|0|1|0|0|0|0|0|0|0|0|0|0|0|0|0|0|0|0|0|0|0|0|0|0|0|0|0|1</HistCounts>
-      </HistItem>
-    </Histograms>
-    <Metadata>
-      <MDI key="STATISTICS_MAXIMUM">10.906113624573</MDI>
-      <MDI key="STATISTICS_MEAN">-0.49696285621855</MDI>
-      <MDI key="STATISTICS_MINIMUM">-10.839770317078</MDI>
-      <MDI key="STATISTICS_STDDEV">1.8890729699247</MDI>
-    </Metadata>
-  </PAMRasterBand>
-</PAMDataset>
