--- Doc/dhdt/Hooker/availableWV.txt
@@ -1,5 +1,0 @@
-11AUG16WV01DEM1
-12JUL20WV01DEM2
-13APR18WV01DEM1
-13APR19WV01DEM3
-14APR02WV02DEM1
