--- Doc/dhdt/Hooker/statistics.txt
@@ -1,30 +1,0 @@
-Name of the WorldView File: 11AUG16WV01DEM1
-Trimming -> Lbound: -1.744668 Ubound: 1.739445
-N        -> Before: 225 After:  194
-Mean     -> -0.025229
-StdDev   -> 0.579438
-
-Name of the WorldView File: 12JUL20WV01DEM2
-Trimming -> Lbound: -1.997720 Ubound: 1.914576
-N        -> Before: 417 After:  361
-Mean     -> 0.003576
-StdDev   -> 0.647260
-
-Name of the WorldView File: 13APR18WV01DEM1
-Trimming -> Lbound: -1.779009 Ubound: 1.731368
-N        -> Before: 224 After:  200
-Mean     -> -0.000750
-StdDev   -> 0.570123
-
-Name of the WorldView File: 13APR19WV01DEM3
-Trimming -> Lbound: -1.514128 Ubound: 1.591511
-N        -> Before: 236 After:  207
-Mean     -> 0.035360
-StdDev   -> 0.551240
-
-Name of the WorldView File: 14APR02WV02DEM1
-Trimming -> Lbound: -1.519404 Ubound: 1.408139
-N        -> Before: 223 After:  199
-Mean     -> -0.042338
-StdDev   -> 0.509941
-
