+++ Doc/dhdt/v0_1/weightedRegression_doc.pdf
